Move more internal state over to input_driver.c and menu_driver.c - retroarch.c now down to 909kb
