+++ retroarch.c
@@ -7034,7 +7034,7 @@ bool command_event(enum event_command cmd, void *data)
          break;
       case CMD_EVENT_OVERLAY_DEINIT:
 #ifdef HAVE_OVERLAY
-         retroarch_overlay_deinit(p_rarch);
+         input_overlay_deinit();
 #endif
 #if defined(HAVE_TRANSLATE) && defined(HAVE_GFX_WIDGETS)
          /* Because the overlay is a display widget,
@@ -7046,7 +7046,7 @@ bool command_event(enum event_command cmd, void *data)
          break;
       case CMD_EVENT_OVERLAY_INIT:
 #ifdef HAVE_OVERLAY
-         retroarch_overlay_init(p_rarch);
+         input_overlay_init();
 #endif
          break;
       case CMD_EVENT_CHEAT_INDEX_PLUS:
@@ -7623,20 +7623,21 @@ bool command_event(enum event_command cmd, void *data)
 #ifdef HAVE_OVERLAY
          {
             bool *check_rotation           = (bool*)data;
+            input_driver_state_t *input_st = input_state_get_ptr();
             bool inp_overlay_auto_rotate   = settings->bools.input_overlay_auto_rotate;
             float input_overlay_opacity    = settings->floats.input_overlay_opacity;
-            if (!p_rarch->overlay_ptr)
+            if (!input_st->overlay_ptr)
                return false;
 
-            p_rarch->overlay_ptr->index    = p_rarch->overlay_ptr->next_index;
-            p_rarch->overlay_ptr->active   = &p_rarch->overlay_ptr->overlays[
-               p_rarch->overlay_ptr->index];
+            input_st->overlay_ptr->index   = input_st->overlay_ptr->next_index;
+            input_st->overlay_ptr->active  = &input_st->overlay_ptr->overlays[
+               input_st->overlay_ptr->index];
 
-            input_overlay_load_active(p_rarch->overlay_visibility,
-                  p_rarch->overlay_ptr, input_overlay_opacity);
+            input_overlay_load_active(input_st->overlay_visibility,
+                  input_st->overlay_ptr, input_overlay_opacity);
 
-            p_rarch->overlay_ptr->blocked    = true;
-            p_rarch->overlay_ptr->next_index = (unsigned)((p_rarch->overlay_ptr->index + 1) % p_rarch->overlay_ptr->size);
+            input_st->overlay_ptr->blocked    = true;
+            input_st->overlay_ptr->next_index = (unsigned)((input_st->overlay_ptr->index + 1) % input_st->overlay_ptr->size);
 
             /* Check orientation, if required */
             if (inp_overlay_auto_rotate)
@@ -7646,7 +7647,7 @@ bool command_event(enum event_command cmd, void *data)
                            p_rarch->video_driver_width,
                            p_rarch->video_driver_height,
                            settings->bools.input_overlay_enable,
-                           p_rarch->overlay_ptr);
+                           input_st->overlay_ptr);
          }
 #endif
          break;
@@ -7897,6 +7898,7 @@ bool command_event(enum event_command cmd, void *data)
 #ifdef HAVE_OVERLAY
          {
             overlay_layout_desc_t layout_desc;
+            input_driver_state_t *input_st      = input_state_get_ptr();
 
             layout_desc.scale_landscape         = settings->floats.input_overlay_scale_landscape;
             layout_desc.aspect_adjust_landscape = settings->floats.input_overlay_aspect_adjust_landscape;
@@ -7913,8 +7915,10 @@ bool command_event(enum event_command cmd, void *data)
             layout_desc.touch_scale             = (float)settings->uints.input_touch_scale;
             layout_desc.auto_scale              = settings->bools.input_overlay_auto_scale;
 
-            input_overlay_set_scale_factor(p_rarch->overlay_ptr, &layout_desc,
-                  p_rarch->video_driver_width, p_rarch->video_driver_height);
+            input_overlay_set_scale_factor(input_st->overlay_ptr,
+                  &layout_desc,
+                  p_rarch->video_driver_width,
+                  p_rarch->video_driver_height);
          }
 #endif
          break;
@@ -7923,9 +7927,11 @@ bool command_event(enum event_command cmd, void *data)
           * The alpha factor is applied for all overlays. */
 #ifdef HAVE_OVERLAY
          {
-            float input_overlay_opacity = settings->floats.input_overlay_opacity;
-            input_overlay_set_alpha_mod(p_rarch->overlay_visibility,
-                  p_rarch->overlay_ptr, input_overlay_opacity);
+            float input_overlay_opacity    = settings->floats.input_overlay_opacity;
+            input_driver_state_t *input_st = input_state_get_ptr();
+
+            input_overlay_set_alpha_mod(input_st->overlay_visibility,
+                  input_st->overlay_ptr, input_overlay_opacity);
          }
 #endif
          break;
@@ -14281,7 +14287,7 @@ static bool video_driver_overlay_interface(
       const video_overlay_interface_t **iface);
 
 /* task_data = overlay_task_data_t* */
-static void input_overlay_loaded(retro_task_t *task,
+void input_overlay_loaded(retro_task_t *task,
       void *task_data, void *user_data, const char *err)
 {
    size_t i;
@@ -14289,10 +14295,11 @@ static void input_overlay_loaded(retro_task_t *task,
    overlay_task_data_t              *data = (overlay_task_data_t*)task_data;
    input_overlay_t                    *ol = NULL;
    const video_overlay_interface_t *iface = NULL;
-   settings_t *settings                   = p_rarch->configuration_settings;
+   settings_t *settings                   = config_get_ptr();
    bool input_overlay_show_mouse_cursor   = settings->bools.input_overlay_show_mouse_cursor;
    bool inp_overlay_auto_rotate           = settings->bools.input_overlay_auto_rotate;
    bool input_overlay_enable              = settings->bools.input_overlay_enable;
+   input_driver_state_t *input_st         = input_state_get_ptr();
    if (err)
       return;
 
@@ -14327,7 +14334,8 @@ static void input_overlay_loaded(retro_task_t *task,
    ol->iface      = iface;
    ol->iface_data = p_rarch->video_driver_data;
 
-   input_overlay_load_active(p_rarch->overlay_visibility, ol, data->overlay_opacity);
+   input_overlay_load_active(input_st->overlay_visibility,
+         ol, data->overlay_opacity);
 
    /* Enable or disable the overlay. */
    ol->enable = data->overlay_enable;
@@ -14345,12 +14353,12 @@ static void input_overlay_loaded(retro_task_t *task,
    /* Due to the asynchronous nature of overlay loading
     * it is possible for overlay_ptr to be non-NULL here
     * > Ensure it is free()'d before assigning new pointer */
-   if (p_rarch->overlay_ptr)
+   if (input_st->overlay_ptr)
    {
-      input_overlay_free_overlays(p_rarch->overlay_ptr);
-      free(p_rarch->overlay_ptr);
+      input_overlay_free_overlays(input_st->overlay_ptr);
+      free(input_st->overlay_ptr);
    }
-   p_rarch->overlay_ptr = ol;
+   input_st->overlay_ptr = ol;
 
    free(data);
 
@@ -14363,7 +14371,7 @@ static void input_overlay_loaded(retro_task_t *task,
             p_rarch->video_driver_width,
             p_rarch->video_driver_height,
             input_overlay_enable,
-            p_rarch->overlay_ptr);
+            input_st->overlay_ptr);
 
    return;
 
@@ -14374,124 +14382,9 @@ abort_load:
    free(data->overlays);
    free(data);
 }
-
-void input_overlay_set_visibility(int overlay_idx,
-      enum overlay_visibility vis)
-{
-   struct rarch_state         *p_rarch = &rarch_st;
-   input_overlay_t                 *ol = p_rarch->overlay_ptr;
-
-   if (!p_rarch->overlay_visibility)
-   {
-      unsigned i;
-      p_rarch->overlay_visibility = (enum overlay_visibility *)calloc(
-            MAX_VISIBILITY, sizeof(enum overlay_visibility));
-
-      for (i = 0; i < MAX_VISIBILITY; i++)
-         p_rarch->overlay_visibility[i] = OVERLAY_VISIBILITY_DEFAULT;
-   }
-
-   p_rarch->overlay_visibility[overlay_idx] = vis;
-
-   if (!ol)
-      return;
-   if (vis == OVERLAY_VISIBILITY_HIDDEN)
-      ol->iface->set_alpha(ol->iface_data, overlay_idx, 0.0);
-}
-
-static void retroarch_overlay_deinit(struct rarch_state *p_rarch)
-{
-   input_overlay_free(p_rarch->overlay_ptr);
-   p_rarch->overlay_ptr = NULL;
-}
-
-static void retroarch_overlay_init(struct rarch_state *p_rarch)
-{
-   settings_t *settings                     = p_rarch->configuration_settings;
-   bool input_overlay_enable                = settings->bools.input_overlay_enable;
-   bool input_overlay_auto_scale            = settings->bools.input_overlay_auto_scale;
-   const char *path_overlay                 = settings->paths.path_overlay;
-   float overlay_opacity                    = settings->floats.input_overlay_opacity;
-   float overlay_scale_landscape            = settings->floats.input_overlay_scale_landscape;
-   float overlay_aspect_adjust_landscape    = settings->floats.input_overlay_aspect_adjust_landscape;
-   float overlay_x_separation_landscape     = settings->floats.input_overlay_x_separation_landscape;
-   float overlay_y_separation_landscape     = settings->floats.input_overlay_y_separation_landscape;
-   float overlay_x_offset_landscape         = settings->floats.input_overlay_x_offset_landscape;
-   float overlay_y_offset_landscape         = settings->floats.input_overlay_y_offset_landscape;
-   float overlay_scale_portrait             = settings->floats.input_overlay_scale_portrait;
-   float overlay_aspect_adjust_portrait     = settings->floats.input_overlay_aspect_adjust_portrait;
-   float overlay_x_separation_portrait      = settings->floats.input_overlay_x_separation_portrait;
-   float overlay_y_separation_portrait      = settings->floats.input_overlay_y_separation_portrait;
-   float overlay_x_offset_portrait          = settings->floats.input_overlay_x_offset_portrait;
-   float overlay_y_offset_portrait          = settings->floats.input_overlay_y_offset_portrait;
-   float overlay_touch_scale                = (float)settings->uints.input_touch_scale;
-
-   bool load_enabled                        = input_overlay_enable;
-#ifdef HAVE_MENU
-   bool overlay_hide_in_menu                = settings->bools.input_overlay_hide_in_menu;
-#else
-   bool overlay_hide_in_menu                = false;
-#endif
-   bool overlay_hide_when_gamepad_connected = settings->bools.input_overlay_hide_when_gamepad_connected;
-#if defined(GEKKO)
-   /* Avoid a crash at startup or even when toggling overlay in rgui */
-   uint64_t memory_free                     = frontend_driver_get_free_memory();
-   if (memory_free < (3 * 1024 * 1024))
-      return;
-#endif
-
-   retroarch_overlay_deinit(p_rarch);
-
-#ifdef HAVE_MENU
-   /* Cancel load if 'hide_in_menu' is enabled and
-    * menu is currently active */
-   if (overlay_hide_in_menu)
-      load_enabled = load_enabled && !menu_state_get_ptr()->alive;
-#endif
-
-   /* Cancel load if 'hide_when_gamepad_connected' is
-    * enabled and a gamepad is currently connected */
-   if (overlay_hide_when_gamepad_connected)
-      load_enabled = load_enabled && (input_config_get_device_name(0) == NULL);
-
-   if (load_enabled)
-   {
-      overlay_layout_desc_t layout_desc;
-
-      layout_desc.scale_landscape         = overlay_scale_landscape;
-      layout_desc.aspect_adjust_landscape = overlay_aspect_adjust_landscape;
-      layout_desc.x_separation_landscape  = overlay_x_separation_landscape;
-      layout_desc.y_separation_landscape  = overlay_y_separation_landscape;
-      layout_desc.x_offset_landscape      = overlay_x_offset_landscape;
-      layout_desc.y_offset_landscape      = overlay_y_offset_landscape;
-      layout_desc.scale_portrait          = overlay_scale_portrait;
-      layout_desc.aspect_adjust_portrait  = overlay_aspect_adjust_portrait;
-      layout_desc.x_separation_portrait   = overlay_x_separation_portrait;
-      layout_desc.y_separation_portrait   = overlay_y_separation_portrait;
-      layout_desc.x_offset_portrait       = overlay_x_offset_portrait;
-      layout_desc.y_offset_portrait       = overlay_y_offset_portrait;
-      layout_desc.touch_scale             = overlay_touch_scale;
-      layout_desc.auto_scale              = input_overlay_auto_scale;
-
-      task_push_overlay_load_default(input_overlay_loaded,
-            path_overlay,
-            overlay_hide_in_menu,
-            overlay_hide_when_gamepad_connected,
-            input_overlay_enable,
-            overlay_opacity,
-            &layout_desc,
-            NULL);
-   }
-}
 #endif
 
 /* INPUT */
-
-void set_connection_listener(pad_connection_listener_t *listener)
-{
-   struct rarch_state *p_rarch      = &rarch_st;
-   p_rarch->pad_connection_listener = listener;
-}
 
 /**
  * config_get_input_driver_options:
@@ -14575,7 +14468,7 @@ static void input_driver_poll(void)
    }
 
 #ifdef HAVE_OVERLAY
-   if (p_rarch->overlay_ptr && p_rarch->overlay_ptr->alive)
+   if (input_st->overlay_ptr && input_st->overlay_ptr->alive)
    {
       unsigned input_analog_dpad_mode = settings->uints.input_analog_dpad_mode[0];
 
@@ -14602,8 +14495,8 @@ static void input_driver_poll(void)
       input_poll_overlay(
             input_st->keyboard_mapping_blocked,
             settings,
-            p_rarch->overlay_ptr,
-            p_rarch->overlay_visibility,
+            input_st->overlay_ptr,
+            input_st->overlay_visibility,
             input_overlay_opacity,
             input_analog_dpad_mode,
             settings->floats.input_axis_threshold);
@@ -14616,8 +14509,8 @@ static void input_driver_poll(void)
    if (input_remap_binds_enable)
    {
 #ifdef HAVE_OVERLAY
-      input_overlay_t *overlay_pointer = (input_overlay_t*)p_rarch->overlay_ptr;
-      bool poll_overlay                = (p_rarch->overlay_ptr && p_rarch->overlay_ptr->alive);
+      input_overlay_t *overlay_pointer = (input_overlay_t*)input_st->overlay_ptr;
+      bool poll_overlay                = (input_st->overlay_ptr && input_st->overlay_ptr->alive);
 #endif
       input_mapper_t *handle           = &input_st->mapper;
       float input_analog_deadzone      = settings->floats.input_analog_deadzone;
@@ -14830,8 +14723,8 @@ static void input_driver_poll(void)
 
 #ifdef HAVE_ACCESSIBILITY
                   /* gamepad override */
-                  if (i == 0 &&
-                        p_rarch->gamepad_input_override & (1 << j))
+                  if (     (i == 0)
+                        && input_st->gamepad_input_override & (1 << j))
                   {
                      BIT256_SET(handle->buttons[i], j);
                   }
@@ -14946,11 +14839,11 @@ static void input_driver_poll(void)
                   sizeof(msg), 0, NULL, NULL);
 
             if (ret == sizeof(msg))
-               input_remote_parse_packet(&p_rarch->remote_st_ptr, &msg, user);
+               input_remote_parse_packet(&input_st->remote_st_ptr, &msg, user);
             else if ((ret != -1) || ((errno != EAGAIN) && (errno != ENOENT)))
 #endif
             {
-               input_remote_state_t *input_state  = &p_rarch->remote_st_ptr;
+               input_remote_state_t *input_state  = &input_st->remote_st_ptr;
                input_state->buttons[user]         = 0;
                input_state->analog[0][user]       = 0;
                input_state->analog[1][user]       = 0;
@@ -14961,353 +14854,6 @@ static void input_driver_poll(void)
       }
    }
 #endif
-}
-
-static int16_t input_state_device(
-      struct rarch_state *p_rarch,
-      input_driver_state_t *input_st,
-      settings_t *settings,
-      input_mapper_t *handle,
-      unsigned input_analog_dpad_mode,
-      int16_t ret,
-      unsigned port, unsigned device,
-      unsigned idx, unsigned id,
-      bool button_mask)
-{
-   int16_t res  = 0;
-
-   switch (device)
-   {
-      case RETRO_DEVICE_JOYPAD:
-
-         if (id < RARCH_FIRST_META_KEY)
-         {
-#ifdef HAVE_NETWORKGAMEPAD
-            /* Don't process binds if input is coming from Remote RetroPad */
-            if (     input_st->remote
-                  && INPUT_REMOTE_KEY_PRESSED(p_rarch, id, port))
-               res |= 1;
-            else
-#endif
-            {
-               bool bind_valid       = input_st->libretro_input_binds[port]
-                  && input_st->libretro_input_binds[port][id].valid;
-               unsigned remap_button = settings->uints.input_remap_ids[port][id];
-
-               /* TODO/FIXME: What on earth is this code doing...? */
-               if (!
-                     (      bind_valid
-                            && id != remap_button
-                     )
-                  )
-               {
-                  if (button_mask)
-                  {
-                     if (ret & (1 << id))
-                        res |= (1 << id);
-                  }
-                  else
-                     res = ret;
-
-               }
-
-               if (BIT256_GET(handle->buttons[port], id))
-                  res = 1;
-
-#ifdef HAVE_OVERLAY
-               /* Check if overlay is active and button
-                * corresponding to 'id' has been pressed */
-               if ((port == 0) &&
-                   p_rarch->overlay_ptr &&
-                   p_rarch->overlay_ptr->alive &&
-                   BIT256_GET(p_rarch->overlay_ptr->overlay_state.buttons, id))
-               {
-#ifdef HAVE_MENU
-                  bool menu_driver_alive        = menu_state_get_ptr()->alive;
-#else
-                  bool menu_driver_alive        = false;
-#endif
-                  bool input_remap_binds_enable = settings->bools.input_remap_binds_enable;
-
-                  /* This button has already been processed
-                   * inside input_driver_poll() if all the
-                   * following are true:
-                   * > Menu driver is not running
-                   * > Input remaps are enabled
-                   * > 'id' is not equal to remapped button index
-                   * If these conditions are met, input here
-                   * is ignored */
-                  if ((menu_driver_alive || !input_remap_binds_enable) ||
-                      (id == remap_button))
-                     res |= 1;
-               }
-#endif
-            }
-
-            /* Don't allow turbo for D-pad. */
-            if (     (id  < RETRO_DEVICE_ID_JOYPAD_UP)    ||
-                  (    (id  > RETRO_DEVICE_ID_JOYPAD_RIGHT) &&
-                       (id <= RETRO_DEVICE_ID_JOYPAD_R3)))
-            {
-               /*
-                * Apply turbo button if activated.
-                */
-               unsigned turbo_mode = settings->uints.input_turbo_mode;
-
-               if (turbo_mode > INPUT_TURBO_MODE_CLASSIC)
-               {
-                  /* Pressing turbo button toggles turbo mode on or off.
-                   * Holding the button will
-                   * pass through, else the pressed state will be modulated by a
-                   * periodic pulse defined by the configured duty cycle.
-                   */
-
-                  /* Avoid detecting the turbo button being held as multiple toggles */
-                  if (!input_st->turbo_btns.frame_enable[port])
-                     input_st->turbo_btns.turbo_pressed[port] &= ~(1 << 31);
-                  else if (input_st->turbo_btns.turbo_pressed[port]>=0)
-                  {
-                     input_st->turbo_btns.turbo_pressed[port] |= (1 << 31);
-                     /* Toggle turbo for selected buttons. */
-                     if (input_st->turbo_btns.enable[port]
-                           != (1 << settings->uints.input_turbo_default_button))
-                     {
-                        static const int button_map[]={
-                           RETRO_DEVICE_ID_JOYPAD_B,
-                           RETRO_DEVICE_ID_JOYPAD_Y,
-                           RETRO_DEVICE_ID_JOYPAD_A,
-                           RETRO_DEVICE_ID_JOYPAD_X,
-                           RETRO_DEVICE_ID_JOYPAD_L,
-                           RETRO_DEVICE_ID_JOYPAD_R,
-                           RETRO_DEVICE_ID_JOYPAD_L2,
-                           RETRO_DEVICE_ID_JOYPAD_R2,
-                           RETRO_DEVICE_ID_JOYPAD_L3,
-                           RETRO_DEVICE_ID_JOYPAD_R3};
-                        input_st->turbo_btns.enable[port] = 1 << button_map[
-                           MIN(
-                                 ARRAY_SIZE(button_map) - 1,
-                                 settings->uints.input_turbo_default_button)];
-                     }
-                     input_st->turbo_btns.mode1_enable[port] ^= 1;
-                  }
-
-                  if (input_st->turbo_btns.turbo_pressed[port] & (1 << 31))
-                  {
-                     /* Avoid detecting buttons being held as multiple toggles */
-                     if (!res)
-                        input_st->turbo_btns.turbo_pressed[port] &= ~(1 << id);
-                     else if (!(input_st->turbo_btns.turbo_pressed[port] & (1 << id)) &&
-                           turbo_mode == INPUT_TURBO_MODE_SINGLEBUTTON)
-                     {
-                        uint16_t enable_new;
-                        input_st->turbo_btns.turbo_pressed[port] |= 1 << id;
-                        /* Toggle turbo for pressed button but make
-                         * sure at least one button has turbo */
-                        enable_new = input_st->turbo_btns.enable[port] ^ (1 << id);
-                        if (enable_new)
-                           input_st->turbo_btns.enable[port] = enable_new;
-                     }
-                  }
-                  else if (turbo_mode == INPUT_TURBO_MODE_SINGLEBUTTON_HOLD &&
-                        input_st->turbo_btns.enable[port] &&
-                        input_st->turbo_btns.mode1_enable[port])
-                  {
-                     /* Hold mode stops turbo on release */
-                     input_st->turbo_btns.mode1_enable[port] = 0;
-                  }
-
-                  if (!res && input_st->turbo_btns.mode1_enable[port] &&
-                        input_st->turbo_btns.enable[port] & (1 << id))
-                  {
-                     /* if turbo button is enabled for this key ID */
-                     res = ((   input_st->turbo_btns.count
-                              % settings->uints.input_turbo_period)
-                           < settings->uints.input_turbo_duty_cycle);
-                  }
-               }
-               else
-               {
-                  /* If turbo button is held, all buttons pressed except
-                   * for D-pad will go into a turbo mode. Until the button is
-                   * released again, the input state will be modulated by a
-                   * periodic pulse defined by the configured duty cycle.
-                   */
-                  if (res)
-                  {
-                     if (input_st->turbo_btns.frame_enable[port])
-                        input_st->turbo_btns.enable[port] |= (1 << id);
-
-                     if (input_st->turbo_btns.enable[port] & (1 << id))
-                        /* if turbo button is enabled for this key ID */
-                        res = ((input_st->turbo_btns.count
-                                 % settings->uints.input_turbo_period)
-                              < settings->uints.input_turbo_duty_cycle);
-                  }
-                  else
-                     input_st->turbo_btns.enable[port] &= ~(1 << id);
-               }
-            }
-         }
-
-         break;
-
-
-      case RETRO_DEVICE_KEYBOARD:
-
-         res = ret;
-
-         if (id < RETROK_LAST)
-         {
-#ifdef HAVE_OVERLAY
-            if (port == 0)
-            {
-               if (p_rarch->overlay_ptr && p_rarch->overlay_ptr->alive)
-               {
-                  input_overlay_state_t
-                     *ol_state          = &p_rarch->overlay_ptr->overlay_state;
-
-                  if (OVERLAY_GET_KEY(ol_state, id))
-                     res               |= 1;
-               }
-            }
-#endif
-            if (MAPPER_GET_KEY(handle, id))
-               res |= 1;
-         }
-
-         break;
-
-
-      case RETRO_DEVICE_ANALOG:
-         {
-#if defined(HAVE_NETWORKGAMEPAD) || defined(HAVE_OVERLAY)
-#ifdef HAVE_NETWORKGAMEPAD
-            input_remote_state_t
-               *input_state         = &p_rarch->remote_st_ptr;
-
-#endif
-            unsigned base           = (idx == RETRO_DEVICE_INDEX_ANALOG_RIGHT)
-               ? 2 : 0;
-            if (id == RETRO_DEVICE_ID_ANALOG_Y)
-               base += 1;
-#ifdef HAVE_NETWORKGAMEPAD
-            if (     input_st->remote
-                  && input_state && input_state->analog[base][port])
-               res          = input_state->analog[base][port];
-            else
-#endif
-#endif
-            {
-               if (id < RARCH_FIRST_META_KEY)
-               {
-                  bool bind_valid         = input_st->libretro_input_binds[port]
-                     && input_st->libretro_input_binds[port][id].valid;
-
-                  if (bind_valid)
-                  {
-                     /* reset_state - used to reset input state of a button
-                      * when the gamepad mapper is in action for that button*/
-                     bool reset_state        = false;
-                     if (idx < 2 && id < 2)
-                     {
-                        unsigned offset = RARCH_FIRST_CUSTOM_BIND +
-                           (idx * 4) + (id * 2);
-
-                        if (settings->uints.input_remap_ids[port][offset] != offset)
-                           reset_state = true;
-                        else if (settings->uints.input_remap_ids[port][offset+1] != (offset+1))
-                           reset_state = true;
-                     }
-
-                     if (reset_state)
-                        res = 0;
-                     else
-                     {
-                        res = ret;
-
-#ifdef HAVE_OVERLAY
-                        if (p_rarch->overlay_ptr &&
-                            p_rarch->overlay_ptr->alive &&
-                            (port == 0) &&
-                            !(((input_analog_dpad_mode == ANALOG_DPAD_LSTICK) &&
-                                 (idx == RETRO_DEVICE_INDEX_ANALOG_LEFT)) ||
-                             ((input_analog_dpad_mode == ANALOG_DPAD_RSTICK) &&
-                                 (idx == RETRO_DEVICE_INDEX_ANALOG_RIGHT))))
-                        {
-                           input_overlay_state_t *ol_state =
-                              &p_rarch->overlay_ptr->overlay_state;
-                           int16_t ol_analog               =
-                                 ol_state->analog[base];
-
-                           /* Analog values are an integer corresponding
-                            * to the extent of the analog motion; these
-                            * cannot be OR'd together, we must instead
-                            * keep the value with the largest magnitude */
-                           if (ol_analog)
-                           {
-                              if (res == 0)
-                                 res = ol_analog;
-                              else
-                              {
-                                 int16_t ol_analog_abs = (ol_analog >= 0) ?
-                                       ol_analog : -ol_analog;
-                                 int16_t res_abs       = (res >= 0) ?
-                                       res : -res;
-
-                                 res = (ol_analog_abs > res_abs) ?
-                                       ol_analog : res;
-                              }
-                           }
-                        }
-#endif
-                     }
-                  }
-               }
-            }
-
-            if (idx < 2 && id < 2)
-            {
-               unsigned offset = 0 + (idx * 4) + (id * 2);
-               int        val1 = handle->analog_value[port][offset];
-               int        val2 = handle->analog_value[port][offset+1];
-
-               /* OR'ing these analog values is 100% incorrect,
-                * but I have no idea what this code is supposed
-                * to be doing (val1 and val2 always seem to be
-                * zero), so I will leave it alone... */
-               if (val1)
-                  res          |= val1;
-               else if (val2)
-                  res          |= val2;
-            }
-         }
-         break;
-
-      case RETRO_DEVICE_MOUSE:
-      case RETRO_DEVICE_LIGHTGUN:
-      case RETRO_DEVICE_POINTER:
-
-         if (id < RARCH_FIRST_META_KEY)
-         {
-            bool bind_valid = input_st->libretro_input_binds[port]
-               && input_st->libretro_input_binds[port][id].valid;
-
-            if (bind_valid)
-            {
-               if (button_mask)
-               {
-                  if (ret & (1 << id))
-                     res |= (1 << id);
-               }
-               else
-                  res = ret;
-            }
-         }
-
-         break;
-   }
-
-   return res;
 }
 
 int16_t input_state_internal(unsigned port, unsigned device,
@@ -15458,14 +15004,14 @@ int16_t input_state_internal(unsigned port, unsigned device,
          {
             unsigned i;
             for (i = 0; i < RARCH_FIRST_CUSTOM_BIND; i++)
-               if (input_state_device(p_rarch, input_st,
+               if (input_state_device(input_st,
                         settings, handle,
                         input_analog_dpad_mode, ret, mapped_port,
                         device, idx, i, true))
                   port_result |= (1 << i);
          }
          else
-            port_result = input_state_device(p_rarch, input_st,
+            port_result = input_state_device(input_st,
                   settings, handle,
                   input_analog_dpad_mode, ret, mapped_port,
                   device, idx, id, false);
@@ -15573,397 +15119,6 @@ static int16_t input_state(unsigned port, unsigned device,
 
 /* MENU INPUT */
 #ifdef HAVE_MENU
-/*
- * This function gets called in order to process all input events
- * for the current frame.
- *
- * Sends input code to menu for one frame.
- *
- * It uses as input the local variables 'input' and 'trigger_input'.
- *
- * Mouse and touch input events get processed inside this function.
- *
- * NOTE: 'input' and 'trigger_input' is sourced from the keyboard and/or
- * the gamepad. It does not contain input state derived from the mouse
- * and/or touch - this gets dealt with separately within this function.
- *
- * TODO/FIXME - maybe needs to be overhauled so we can send multiple
- * events per frame if we want to, and we shouldn't send the
- * entire button state either but do a separate event per button
- * state.
- */
-static unsigned menu_event(
-      struct rarch_state *p_rarch,
-      settings_t *settings,
-      input_bits_t *p_input,
-      input_bits_t *p_trigger_input,
-      bool display_kb)
-{
-   /* Used for key repeat */
-   static float delay_timer                        = 0.0f;
-   static float delay_count                        = 0.0f;
-   static bool initial_held                        = true;
-   static bool first_held                          = false;
-   static unsigned ok_old                          = 0;
-   unsigned ret                                    = MENU_ACTION_NOOP;
-   bool set_scroll                                 = false;
-   size_t new_scroll_accel                         = 0;
-   struct menu_state                     *menu_st  = menu_state_get_ptr();
-   menu_input_t *menu_input                        = &menu_st->input_state;
-   input_driver_state_t *input_st                  = input_state_get_ptr();
-   input_driver_t *current_input                   = input_st->current_driver;
-   const input_device_driver_t
-      *joypad                                      = input_st->primary_joypad;
-#ifdef HAVE_MFI
-   const input_device_driver_t *sec_joypad         =
-      input_st->secondary_joypad;
-#else
-   const input_device_driver_t *sec_joypad         = NULL;
-#endif
-   gfx_display_t *p_disp                           = disp_get_ptr();
-   menu_input_pointer_hw_state_t *pointer_hw_state = &menu_st->input_pointer_hw_state;
-   menu_handle_t             *menu                 = menu_st->driver_data;
-   bool keyboard_mapping_blocked                   = input_st->keyboard_mapping_blocked;
-   bool menu_mouse_enable                          = settings->bools.menu_mouse_enable;
-   bool menu_pointer_enable                        = settings->bools.menu_pointer_enable;
-   bool swap_ok_cancel_btns                        = settings->bools.input_menu_swap_ok_cancel_buttons;
-   bool menu_scroll_fast                           = settings->bools.menu_scroll_fast;
-   bool pointer_enabled                            = settings->bools.menu_pointer_enable;
-   unsigned input_touch_scale                      = settings->uints.input_touch_scale;
-   unsigned menu_scroll_delay                      =
-      settings->uints.menu_scroll_delay;
-#ifdef HAVE_OVERLAY
-   bool input_overlay_enable                       = settings->bools.input_overlay_enable;
-   bool overlay_active                             = input_overlay_enable && p_rarch->overlay_ptr
-      && p_rarch->overlay_ptr->alive;
-#else
-   bool input_overlay_enable                       = false;
-   bool overlay_active                             = false;
-#endif
-   unsigned menu_ok_btn                            = swap_ok_cancel_btns ?
-         RETRO_DEVICE_ID_JOYPAD_B : RETRO_DEVICE_ID_JOYPAD_A;
-   unsigned menu_cancel_btn                        = swap_ok_cancel_btns ?
-         RETRO_DEVICE_ID_JOYPAD_A : RETRO_DEVICE_ID_JOYPAD_B;
-   unsigned ok_current                             = BIT256_GET_PTR(p_input, menu_ok_btn);
-   unsigned ok_trigger                             = ok_current & ~ok_old;
-   unsigned i                                      = 0;
-   static unsigned navigation_initial              = 0;
-   unsigned navigation_current                     = 0;
-   unsigned navigation_buttons[6] =
-   {
-      RETRO_DEVICE_ID_JOYPAD_UP,
-      RETRO_DEVICE_ID_JOYPAD_DOWN,
-      RETRO_DEVICE_ID_JOYPAD_LEFT,
-      RETRO_DEVICE_ID_JOYPAD_RIGHT,
-      RETRO_DEVICE_ID_JOYPAD_L,
-      RETRO_DEVICE_ID_JOYPAD_R
-   };
-
-   ok_old                                          = ok_current;
-
-   /* Get pointer (mouse + touchscreen) input
-    * Note: Must be done regardless of menu screensaver
-    *       state */
-
-   /* > If pointer input is disabled, do nothing */
-   if (!menu_mouse_enable && !menu_pointer_enable)
-      menu_input->pointer.type = MENU_POINTER_DISABLED;
-   else
-   {
-      menu_input_pointer_hw_state_t mouse_hw_state       = {0};
-      menu_input_pointer_hw_state_t touchscreen_hw_state = {0};
-
-      /* Read mouse */
-      if (menu_mouse_enable)
-         menu_input_get_mouse_hw_state(
-               p_disp,
-               menu,
-               input_st,
-               current_input,
-               joypad,
-               sec_joypad,
-               keyboard_mapping_blocked,
-               menu_mouse_enable,
-               input_overlay_enable,
-               overlay_active,
-               &mouse_hw_state);
-
-      /* Read touchscreen
-       * Note: Could forgo this if mouse is currently active,
-       * but this is 'cleaner' code... (if performance is a
-       * concern - and it isn't - user can just disable touch
-       * screen support) */
-      if (menu_pointer_enable)
-         menu_input_get_touchscreen_hw_state(
-               p_disp,
-               menu,
-               input_st,
-               current_input,
-               joypad,
-               sec_joypad,
-               keyboard_mapping_blocked,
-               overlay_active,
-               pointer_enabled,
-               input_touch_scale,
-               &touchscreen_hw_state);
-
-      /* Mouse takes precedence */
-      if (mouse_hw_state.active)
-         menu_input->pointer.type = MENU_POINTER_MOUSE;
-      else if (touchscreen_hw_state.active)
-         menu_input->pointer.type = MENU_POINTER_TOUCHSCREEN;
-
-      /* Copy input from the current device */
-      if (menu_input->pointer.type == MENU_POINTER_MOUSE)
-         memcpy(pointer_hw_state, &mouse_hw_state, sizeof(menu_input_pointer_hw_state_t));
-      else if (menu_input->pointer.type == MENU_POINTER_TOUCHSCREEN)
-         memcpy(pointer_hw_state, &touchscreen_hw_state, sizeof(menu_input_pointer_hw_state_t));
-
-      if (pointer_hw_state->active)
-         menu_st->input_last_time_us = menu_st->current_time_us;
-   }
-
-   /* Populate menu_input_state
-    * Note: dx, dy, ptr, y_accel, etc. entries are set elsewhere */
-   menu_input->pointer.x          = pointer_hw_state->x;
-   menu_input->pointer.y          = pointer_hw_state->y;
-   if (menu_input->select_inhibit || menu_input->cancel_inhibit)
-   {
-      menu_input->pointer.active  = false;
-      menu_input->pointer.pressed = false;
-   }
-   else
-   {
-      menu_input->pointer.active  = pointer_hw_state->active;
-      menu_input->pointer.pressed = pointer_hw_state->select_pressed;
-   }
-
-   /* If menu screensaver is active, any input
-    * is intercepted and used to switch it off */
-   if (menu_st->screensaver_active)
-   {
-      /* Check pointer input */
-      bool input_active = (menu_input->pointer.type != MENU_POINTER_DISABLED) &&
-            menu_input->pointer.active;
-
-      /* Check regular input */
-      if (!input_active)
-         input_active = bits_any_set(p_input->data, ARRAY_SIZE(p_input->data));
-
-      if (!input_active)
-         input_active = bits_any_set(p_trigger_input->data, ARRAY_SIZE(p_trigger_input->data));
-
-      /* Disable screensaver if required */
-      if (input_active)
-      {
-         menu_ctx_environment_t menu_environ;
-         menu_environ.type           = MENU_ENVIRON_DISABLE_SCREENSAVER;
-         menu_environ.data           = NULL;
-         menu_st->screensaver_active = false;
-         menu_st->input_last_time_us = menu_st->current_time_us;
-         menu_driver_ctl(RARCH_MENU_CTL_ENVIRONMENT, &menu_environ);
-      }
-
-      /* Annul received input */
-      menu_input->pointer.active      = false;
-      menu_input->pointer.pressed     = false;
-      menu_input->select_inhibit      = true;
-      menu_input->cancel_inhibit      = true;
-      pointer_hw_state->up_pressed    = false;
-      pointer_hw_state->down_pressed  = false;
-      pointer_hw_state->left_pressed  = false;
-      pointer_hw_state->right_pressed = false;
-      return MENU_ACTION_NOOP;
-   }
-
-   /* Accelerate only navigation buttons */
-   for (i = 0; i < 6; i++)
-   {
-      if (BIT256_GET_PTR(p_input, navigation_buttons[i]))
-         navigation_current        |= (1 << navigation_buttons[i]);
-   }
-
-   if (navigation_current)
-   {
-      if (!first_held)
-      {
-         /* Store first direction in order to block "diagonals" */
-         if (!navigation_initial)
-            navigation_initial      = navigation_current;
-
-         /* don't run anything first frame, only capture held inputs
-          * for old_input_state. */
-
-         first_held                 = true;
-         if (initial_held)
-            delay_timer             = menu_scroll_delay;
-         else
-            delay_timer             = menu_scroll_fast ? 100 : 20;
-         delay_count                = 0;
-      }
-
-      if (delay_count >= delay_timer)
-      {
-         uint32_t input_repeat      = 0;
-         for (i = 0; i < 6; i++)
-            BIT32_SET(input_repeat, navigation_buttons[i]);
-
-         set_scroll                 = true;
-         first_held                 = false;
-         p_trigger_input->data[0]  |= p_input->data[0] & input_repeat;
-         new_scroll_accel           = menu_st->scroll.acceleration;
-
-         if (menu_scroll_fast)
-            new_scroll_accel        = MIN(new_scroll_accel + 1, 64);
-         else
-            new_scroll_accel        = MIN(new_scroll_accel + 1, 5);
-      }
-
-      initial_held                  = false;
-   }
-   else
-   {
-      set_scroll                    = true;
-      first_held                    = false;
-      initial_held                  = true;
-      navigation_initial            = 0;
-   }
-
-   if (set_scroll)
-      menu_st->scroll.acceleration  = (unsigned)(new_scroll_accel);
-
-   delay_count                     += anim_get_ptr()->delta_time;
-
-   if (display_kb)
-   {
-      bool show_osk_symbols = input_event_osk_show_symbol_pages(menu_st->driver_data);
-
-      input_event_osk_iterate(input_st->osk_grid, input_st->osk_idx);
-
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_DOWN))
-      {
-         menu_st->input_last_time_us = menu_st->current_time_us;
-         if (input_st->osk_ptr < 33)
-            input_st->osk_ptr += OSK_CHARS_PER_LINE;
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_UP))
-      {
-         menu_st->input_last_time_us = menu_st->current_time_us;
-         if (input_st->osk_ptr >= OSK_CHARS_PER_LINE)
-            input_st->osk_ptr -= OSK_CHARS_PER_LINE;
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_RIGHT))
-      {
-         menu_st->input_last_time_us = menu_st->current_time_us;
-         if (input_st->osk_ptr < 43)
-            input_st->osk_ptr += 1;
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_LEFT))
-      {
-         menu_st->input_last_time_us = menu_st->current_time_us;
-         if (input_st->osk_ptr >= 1)
-            input_st->osk_ptr -= 1;
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_L))
-      {
-         menu_st->input_last_time_us = menu_st->current_time_us;
-         if (input_st->osk_idx > OSK_TYPE_UNKNOWN + 1)
-            input_st->osk_idx = ((enum osk_type)
-                  (input_st->osk_idx - 1));
-         else
-            input_st->osk_idx = ((enum osk_type)(show_osk_symbols
-                     ? OSK_TYPE_LAST - 1
-                     : OSK_SYMBOLS_PAGE1));
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_R))
-      {
-         menu_st->input_last_time_us = menu_st->current_time_us;
-         if (input_st->osk_idx < (show_osk_symbols
-                  ? OSK_TYPE_LAST - 1
-                  : OSK_SYMBOLS_PAGE1))
-            input_st->osk_idx = ((enum osk_type)(
-                     input_st->osk_idx + 1));
-         else
-            input_st->osk_idx = ((enum osk_type)(OSK_TYPE_UNKNOWN + 1));
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, menu_ok_btn))
-      {
-         if (input_st->osk_ptr >= 0)
-            input_event_osk_append(
-                  &input_st->keyboard_line,
-                  &input_st->osk_idx,
-                  &input_st->osk_last_codepoint,
-                  &input_st->osk_last_codepoint_len,
-                  input_st->osk_ptr,
-                  show_osk_symbols,
-                  input_st->osk_grid[input_st->osk_ptr]);
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, menu_cancel_btn))
-         input_keyboard_event(true, '\x7f', '\x7f',
-               0, RETRO_DEVICE_KEYBOARD);
-
-      /* send return key to close keyboard input window */
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_START))
-         input_keyboard_event(true, '\n', '\n', 0, RETRO_DEVICE_KEYBOARD);
-
-      BIT256_CLEAR_ALL_PTR(p_trigger_input);
-   }
-   else
-   {
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_UP))
-      {
-         if (navigation_initial == (1 << RETRO_DEVICE_ID_JOYPAD_UP))
-            ret = MENU_ACTION_UP;
-      }
-      else if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_DOWN))
-      {
-         if (navigation_initial == (1 << RETRO_DEVICE_ID_JOYPAD_DOWN))
-            ret = MENU_ACTION_DOWN;
-      }
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_LEFT))
-      {
-         if (navigation_initial == (1 << RETRO_DEVICE_ID_JOYPAD_LEFT))
-            ret = MENU_ACTION_LEFT;
-      }
-      else if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_RIGHT))
-      {
-         if (navigation_initial == (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT))
-            ret = MENU_ACTION_RIGHT;
-      }
-
-      if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_L))
-         ret = MENU_ACTION_SCROLL_UP;
-      else if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_R))
-         ret = MENU_ACTION_SCROLL_DOWN;
-      else if (ok_trigger)
-         ret = MENU_ACTION_OK;
-      else if (BIT256_GET_PTR(p_trigger_input, menu_cancel_btn))
-         ret = MENU_ACTION_CANCEL;
-      else if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_X))
-         ret = MENU_ACTION_SEARCH;
-      else if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_Y))
-         ret = MENU_ACTION_SCAN;
-      else if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_START))
-         ret = MENU_ACTION_START;
-      else if (BIT256_GET_PTR(p_trigger_input, RETRO_DEVICE_ID_JOYPAD_SELECT))
-         ret = MENU_ACTION_INFO;
-      else if (BIT256_GET_PTR(p_trigger_input, RARCH_MENU_TOGGLE))
-         ret = MENU_ACTION_TOGGLE;
-
-      if (ret != MENU_ACTION_NOOP)
-         menu_st->input_last_time_us = menu_st->current_time_us;
-   }
-
-   return ret;
-}
-
 static int menu_input_pointer_post_iterate(
       struct rarch_state *p_rarch,
       gfx_display_t *p_disp,
@@ -16581,37 +15736,6 @@ static int menu_input_post_iterate(
 }
 #endif
 
-static INLINE bool input_keys_pressed_other_sources(
-      struct rarch_state *p_rarch,
-      input_driver_state_t *input_st,
-      unsigned i,
-      input_bits_t* p_new_state)
-{
-#ifdef HAVE_COMMAND
-   int j;
-   for (j = 0; j < ARRAY_SIZE(input_st->command); j++)
-      if ((i < RARCH_BIND_LIST_END) && input_st->command[j]
-         && input_st->command[j]->state[i])
-         return true;
-#endif
-
-#ifdef HAVE_OVERLAY
-   if (p_rarch->overlay_ptr &&
-         ((BIT256_GET(p_rarch->overlay_ptr->overlay_state.buttons, i))))
-      return true;
-#endif
-
-#ifdef HAVE_NETWORKGAMEPAD
-   /* Only process key presses related to game input if using Remote RetroPad */
-   if (i < RARCH_CUSTOM_BIND_LIST_END
-         && input_st->remote
-         && INPUT_REMOTE_KEY_PRESSED(p_rarch, i, 0))
-      return true;
-#endif
-
-   return false;
-}
-
 /**
  * input_keys_pressed:
  *
@@ -16706,7 +15830,7 @@ static void input_keys_pressed(
       {
          if (
                (ret & (UINT64_C(1) <<  i)) ||
-               input_keys_pressed_other_sources(p_rarch, input_st,
+               input_keys_pressed_other_sources(input_st,
                   i, p_new_state))
          {
             BIT256_SET_PTR(p_new_state, i);
@@ -16721,7 +15845,7 @@ static void input_keys_pressed(
       {
          if (
                   BIT64_GET(lifecycle_state, i)
-               || input_keys_pressed_other_sources(p_rarch, input_st,
+               || input_keys_pressed_other_sources(input_st,
                   i, p_new_state))
          {
             BIT256_SET_PTR(p_new_state, i);
@@ -16744,7 +15868,7 @@ static void input_keys_pressed(
                   port, RETRO_DEVICE_JOYPAD, 0, i);
          if (     bit_pressed
                || BIT64_GET(lifecycle_state, i)
-               || input_keys_pressed_other_sources(p_rarch, input_st,
+               || input_keys_pressed_other_sources(input_st,
                   i, p_new_state))
          {
             BIT256_SET_PTR(p_new_state, i);
@@ -16763,22 +15887,6 @@ static void input_keys_pressed(
 const char* config_get_joypad_driver_options(void)
 {
    return char_list_new_special(STRING_LIST_INPUT_JOYPAD_DRIVERS, NULL);
-}
-
-void input_pad_connect(unsigned port, input_device_driver_t *driver)
-{
-   struct rarch_state *p_rarch = &rarch_st;
-   if (port >= MAX_USERS || !driver)
-   {
-      RARCH_ERR("[Input]: input_pad_connect: bad parameters\n");
-      return;
-   }
-
-   if (p_rarch->pad_connection_listener)
-      p_rarch->pad_connection_listener->connected(port, driver);
-
-   input_autoconfigure_connect(driver->name(port), NULL, driver->ident,
-          port, 0, 0);
 }
 
 #ifdef HAVE_HID
@@ -19724,8 +18832,8 @@ static bool video_driver_init_internal(
          retroarch_fail(p_rarch, 1, "video_driver_init_input()");
 
 #ifdef HAVE_OVERLAY
-   retroarch_overlay_deinit(p_rarch);
-   retroarch_overlay_init(p_rarch);
+   input_overlay_deinit();
+   input_overlay_init();
 #endif
 
 #ifdef HAVE_VIDEO_LAYOUT
@@ -24870,8 +23978,8 @@ void retroarch_menu_running(void)
             settings,
             menu_st->alive,
 #ifdef HAVE_OVERLAY
-            p_rarch->overlay_ptr &&
-            p_rarch->overlay_ptr->alive,
+            input_st->overlay_ptr &&
+            input_st->overlay_ptr->alive,
 #else
             false,
 #endif
@@ -24942,8 +24050,8 @@ void retroarch_menu_running_finished(bool quit)
             settings,
             menu_st->alive,
 #ifdef HAVE_OVERLAY
-            p_rarch->overlay_ptr &&
-            p_rarch->overlay_ptr->alive,
+            input_st->overlay_ptr &&
+            input_st->overlay_ptr->alive,
 #else
             false,
 #endif
@@ -24999,7 +24107,7 @@ void retroarch_menu_running_finished(bool quit)
 #ifdef HAVE_OVERLAY
    if (!quit)
       if (settings && settings->bools.input_overlay_hide_in_menu)
-         retroarch_overlay_init(p_rarch);
+         input_overlay_init();
 #endif
 }
 
@@ -26095,13 +25203,13 @@ static enum runloop_state runloop_check_state(
          {
             unsigned i;
 
-            p_rarch->gamepad_input_override = 0;
+            input_st->gamepad_input_override = 0;
 
             for (i = 0; i < MAX_USERS; i++)
             {
                /* Set gamepad input override */
                if (p_rarch->ai_gamepad_state[i] == 2)
-                  p_rarch->gamepad_input_override |= (1 << i);
+                  input_st->gamepad_input_override |= (1 << i);
                p_rarch->ai_gamepad_state[i] = 0;
             }
          }
@@ -26204,9 +25312,9 @@ static enum runloop_state runloop_check_state(
          if (controller_connected != last_controller_connected)
          {
             if (controller_connected)
-               retroarch_overlay_deinit(p_rarch);
+               input_overlay_deinit();
             else
-               retroarch_overlay_init(p_rarch);
+               input_overlay_init();
 
             last_controller_connected = controller_connected;
          }
@@ -26221,7 +25329,7 @@ static enum runloop_state runloop_check_state(
       else if (prev_overlay_restore)
       {
          if (!input_overlay_hide_in_menu)
-            retroarch_overlay_init(p_rarch);
+            input_overlay_init();
          prev_overlay_restore = false;
       }
 
@@ -26238,7 +25346,7 @@ static enum runloop_state runloop_check_state(
                   p_rarch->video_driver_width,
                   p_rarch->video_driver_height,
                   settings->bools.input_overlay_enable,
-                  p_rarch->overlay_ptr);
+                  input_st->overlay_ptr);
 
          last_width  = video_driver_width;
          last_height = video_driver_height;
@@ -26436,7 +25544,7 @@ static enum runloop_state runloop_check_state(
       bits_clear_bits(trigger_input.data, old_input.data,
             ARRAY_SIZE(trigger_input.data));
       action                    = (enum menu_action)menu_event(
-            p_rarch, settings,
+            settings,
             &current_bits, &trigger_input, display_kb);
       focused                   = pause_nonactive ? is_focused : true;
       focused                   = focused &&
